backend/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
pydantic
rapidfuzz
spacy
metaphone
